re-added base input as it was and renamed new component as textfield
