+++ src/components/atoms/index.ts
@@ -2,3 +2,4 @@ export { default as BaseButton } from './BaseButton.atom.vue'
 export { default as BaseInput } from './BaseInput.atom.vue'
 export { default as BaseText } from './BaseText.atom.vue'
 export { default as BaseIcon } from './BaseIcon.atom.vue'
+export { default as TextField } from './TextField.atom.vue'
